# .migration-backup/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bluesky Feed Generator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Feed {
  id: number;
  recordName: string;
  displayName: string;
  /** @nullable */
  description?: string | null;
  isActive: boolean;
  /** @nullable */
  publishedAt?: string | null;
  postCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateFeedBody {
  recordName: string;
  displayName: string;
  /** @nullable */
  description?: string | null;
}

export interface UpdateFeedBody {
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  isActive?: boolean | null;
}

export interface Keyword {
  id: number;
  feedId: number;
  keyword: string;
  createdAt: string;
}

export interface CreateKeywordBody {
  keyword: string;
}

export interface IndexedPost {
  id: number;
  uri: string;
  cid: string;
  author: string;
  text: string;
  algoTags: string;
  indexedAt: string;
  likes: number;
  reposts: number;
  replies: number;
  quotes: number;
  /** @nullable */
  engagementSyncedAt?: string | null;
}

export type RankedPost = IndexedPost & {
  /** @nullable */
  rank?: number | null;
  /** @nullable */
  finalScore?: number | null;
  /** @nullable */
  qualityScore?: number | null;
  /** @nullable */
  computedAt?: string | null;
};

export interface PostsPage {
  posts: RankedPost[];
  /** @nullable */
  cursor?: string | null;
  total: number;
  mode?: string;
}

export interface StatsOverview {
  totalFeeds: number;
  activeFeeds: number;
  totalPosts: number;
  postsLast24h: number;
  postsLast1h: number;
  firehoseConnected: boolean;
  uptime: number;
}

export interface FirehoseStatus {
  connected: boolean;
  endpoint: string;
  reconnectCount: number;
  /** @nullable */
  lastEventAt?: string | null;
  postsIndexedTotal: number;
}

export interface ActivityBucket {
  hour: string;
  count: number;
}

export interface PublishFeedResult {
  uri: string;
  cid: string;
  feedUri: string;
}

export interface FeedRanking {
  feedId: number;
  recordName: string;
  displayName: string;
  postCount: number;
}

export interface DailyBucket {
  day: string;
  count: number;
}

export interface KeywordStat {
  keyword: string;
  postCount: number;
  percentage: number;
}

export interface AuthorStat {
  did: string;
  postCount: number;
  latestPostAt: string;
}

export interface BlueskyProfile {
  did: string;
  handle: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  description?: string | null;
  followersCount: number;
  followsCount: number;
  postsCount: number;
}

export interface BlueskyFeedInfo {
  uri: string;
  cid: string;
  displayName: string;
  /** @nullable */
  description?: string | null;
  likeCount: number;
  /** @nullable */
  viewerLiked?: string | null;
}

export interface AudienceUser {
  did: string;
  handle: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  description?: string | null;
  followersCount: number;
  followsCount: number;
  /** @nullable */
  followedAt?: string | null;
  /** @nullable */
  followUri?: string | null;
}

export interface AudiencePage {
  users: AudienceUser[];
  /** @nullable */
  cursor?: string | null;
  total: number;
}

export interface SyncResult {
  updated: number;
  skipped: number;
  errors: number;
}

export interface BulkActionBody {
  dids: string[];
  followUris?: string[];
}

export interface BulkActionResult {
  succeeded: number;
  failed: number;
  errors: string[];
}

export interface TopPost {
  id: number;
  uri: string;
  cid: string;
  author: string;
  text: string;
  algoTags: string;
  indexedAt: string;
  likes: number;
  reposts: number;
  replies: number;
  quotes: number;
  totalEngagement: number;
  /** @nullable */
  engagementSyncedAt?: string | null;
}

export interface EngagementOverview {
  totalPosts: number;
  syncedPosts: number;
  totalLikes: number;
  totalReposts: number;
  totalReplies: number;
  totalQuotes: number;
  totalEngagement: number;
  avgLikesPerPost: number;
  /** @nullable */
  topPostUri?: string | null;
}

export interface MyPost {
  uri: string;
  cid: string;
  text: string;
  createdAt: string;
  indexedAt: string;
  likes: number;
  reposts: number;
  replies: number;
  quotes: number;
  hasImages: boolean;
  langs: string[];
}

export interface MyPostStats {
  totalLikes: number;
  totalReposts: number;
  totalReplies: number;
  totalQuotes: number;
  postCount: number;
}

export interface MyPostsResponse {
  posts: MyPost[];
  /** @nullable */
  cursor?: string | null;
  stats: MyPostStats;
}

export interface ComposeBody {
  text: string;
  threadParts?: string[];
  isThread?: boolean;
}

export interface ComposeResult {
  uri: string;
  cid: string;
  uris: string[];
}

export interface ScheduledPost {
  id: number;
  text: string;
  /** @nullable */
  threadParts?: string | null;
  isThread: boolean;
  scheduledAt: string;
  /** @nullable */
  sentAt?: string | null;
  status: string;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
}

export interface CreateScheduledPostBody {
  text: string;
  threadParts?: string[];
  isThread?: boolean;
  scheduledAt: string;
}

export type HashtagStatTopPost = {
  text: string;
  likes: number;
  reposts: number;
} | null;

export interface HashtagStat {
  tag: string;
  postCount: number;
  totalLikes: number;
  totalReposts: number;
  totalReplies: number;
  totalEngagement: number;
  avgEngagement: number;
  avgLikes: number;
  topPost?: HashtagStatTopPost;
}

export interface HashtagAnalysisResponse {
  hashtags: HashtagStat[];
  totalPostsAnalyzed: number;
  postsWithHashtags: number;
}

export interface BestTimeSlot {
  hour: number;
  dayOfWeek: number;
  avgLikes: number;
  avgReposts: number;
  avgReplies: number;
  postCount: number;
  avgEngagement: number;
}

export interface BestTimeResponse {
  hourly: BestTimeSlot[];
  bestHour: number;
  bestDay: number;
}

export interface FollowerSnapshot {
  id: number;
  followersCount: number;
  followsCount: number;
  postsCount: number;
  recordedAt: string;
}

export type GetFeedPostsParams = {
  limit?: number;
  /**
   * @nullable
   */
  cursor?: string | null;
  mode?: GetFeedPostsMode;
};

export type GetFeedPostsMode =
  (typeof GetFeedPostsMode)[keyof typeof GetFeedPostsMode];

export const GetFeedPostsMode = {
  recent: "recent",
  ranked: "ranked",
} as const;

export type ListPostsParams = {
  limit?: number;
  /**
   * @nullable
   */
  cursor?: string | null;
  /**
   * @nullable
   */
  search?: string | null;
};

export type GetTopPostsParams = {
  /**
   * @nullable
   */
  feedId?: number | null;
  limit?: number;
  sortBy?: GetTopPostsSortBy;
};

export type GetTopPostsSortBy =
  (typeof GetTopPostsSortBy)[keyof typeof GetTopPostsSortBy];

export const GetTopPostsSortBy = {
  total: "total",
  likes: "likes",
  reposts: "reposts",
  replies: "replies",
} as const;

export type GetEngagementOverviewParams = {
  /**
   * @nullable
   */
  feedId?: number | null;
};

export type SyncEngagementBody = {
  /** @nullable */
  feedId?: number | null;
  limit?: number;
};

export type GetFollowersParams = {
  /**
   * @nullable
   */
  cursor?: string | null;
  limit?: number;
};

export type GetFollowingParams = {
  /**
   * @nullable
   */
  cursor?: string | null;
  limit?: number;
};

export type GetMyPostsParams = {
  /**
   * @nullable
   */
  cursor?: string | null;
  limit?: number;
};

export type DeleteScheduledPost200 = {
  ok: boolean;
};