# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bluesky Feed Generator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /feeds:
    get:
      operationId: listFeeds
      tags: [feeds]
      summary: List all feeds
      responses:
        "200":
          description: List of feeds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Feed"
    post:
      operationId: createFeed
      tags: [feeds]
      summary: Create a new feed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeedBody"
      responses:
        "201":
          description: Created feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feed"
        "400":
          description: Bad request
  /feeds/{id}:
    get:
      operationId: getFeed
      tags: [feeds]
      summary: Get a single feed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feed"
        "404":
          description: Not found
    patch:
      operationId: updateFeed
      tags: [feeds]
      summary: Update a feed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFeedBody"
      responses:
        "200":
          description: Updated feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feed"
        "404":
          description: Not found
    delete:
      operationId: deleteFeed
      tags: [feeds]
      summary: Delete a feed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /feeds/{id}/keywords:
    get:
      operationId: getFeedKeywords
      tags: [feeds]
      summary: Get keywords for a feed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Keywords
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Keyword"
    post:
      operationId: addFeedKeyword
      tags: [feeds]
      summary: Add a keyword to a feed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKeywordBody"
      responses:
        "201":
          description: Created keyword
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Keyword"
  /feeds/{id}/keywords/{keywordId}:
    delete:
      operationId: deleteFeedKeyword
      tags: [feeds]
      summary: Remove a keyword from a feed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: keywordId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /feeds/{id}/posts:
    get:
      operationId: getFeedPosts
      tags: [posts]
      summary: Get indexed posts for a feed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: cursor
          in: query
          schema:
            type: ["string", "null"]
        - name: mode
          in: query
          schema:
            type: string
            enum: [recent, ranked]
            default: recent
      responses:
        "200":
          description: Posts page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsPage"
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List all indexed posts (across all feeds)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: cursor
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Posts page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsPage"
...
components:
  schemas:
...
    Feed:
      type: object
      properties:
        id:
          type: integer
        recordName:
          type: string
        displayName:
          type: string
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
        publishedAt:
          type: ["string", "null"]
        postCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - recordName
        - displayName
        - isActive
        - postCount
        - createdAt
        - updatedAt
    CreateFeedBody:
      type: object
      properties:
        recordName:
          type: string
        displayName:
          type: string
        description:
          type: ["string", "null"]
      required:
        - recordName
        - displayName
    UpdateFeedBody:
      type: object
      properties:
        displayName:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
    Keyword:
      type: object
      properties:
        id:
          type: integer
        feedId:
          type: integer
        keyword:
          type: string
        createdAt:
          type: string
      required:
        - id
        - feedId
        - keyword
        - createdAt
    CreateKeywordBody:
      type: object
      properties:
        keyword:
          type: string
      required:
        - keyword
    IndexedPost:
      type: object
      properties:
        id:
          type: integer
        uri:
          type: string
        cid:
          type: string
        author:
          type: string
        text:
          type: string
        algoTags:
          type: string
        indexedAt:
          type: string
        likes:
          type: integer
        reposts:
          type: integer
        replies:
          type: integer
        quotes:
          type: integer
        engagementSyncedAt:
          type: ["string", "null"]
      required:
        - id
        - uri
        - cid
        - author
        - text
        - algoTags
        - indexedAt
        - likes
        - reposts
        - replies
        - quotes
    RankedPost:
      allOf:
        - $ref: "#/components/schemas/IndexedPost"
        - type: object
          properties:
            rank:
              type: ["integer", "null"]
            finalScore:
              type: ["number", "null"]
            qualityScore:
              type: ["number", "null"]
            computedAt:
              type: ["string", "null"]
    PostsPage:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/RankedPost"
        cursor:
          type: ["string", "null"]
        total:
          type: integer
        mode:
          type: string
      required:
        - posts
        - total
...